Set value given nil
Because these values are going to be coming from a JSON API that I don't
trust, I'm not entirely sure if they will be returning `undefined` or
`null` when the value doesn't exist, or if it will be consistent.

Because I want this to be consistent, I'm going to make sure that the
default will always work, no matter what.