Move main entities into entities namespace
These are the things that are going to be returned from our
repositories, and this is even how the musicbrainz site defines them, so
I'm going to organize this in a similar matter.

I'm not moving Entity itself, because I might use that for things that
are not explicitly data entities in terms of something from the
musicbrainz API, but something else entirely.