Add basic configuration object
When doing requests to the musicbrainz API we need to give them a
detailed User-Agent so that we don't get as throttled doing requests.

To do this someone can call the configure method, provide a block that
will receive the config object in which they can set the app_name and
contact info.

You can also set the base uri for the API if the gem is to be used with
a different instance of the musicbrainz API.