Add Request object
This is a very tiny wrapper around Typhoeus (it even returns a Typhoeus
response for now. The only thing it's really for is to allow for a
central location to inject the basic information into so that nothing
else should have to worry about attempting to setup headers, base urls
or anything else.