Add IdSearch options
These are the things that we need to specify when doing a search for
something by id. We obviously need the id, and if there are other
relationships we want returned (to other entities) they can be specified
using the relationships array. This will result in more data being
returned for those.