Add base class for Repository
I actually built this as just the artist repository first. But looking
at the code I had, it looked like a bunch of the code would be the same
among a bunch of the different repositories (at least for id search) so
I decided to make a base class. Maybe this would be better as a mixin,
but I think it's fine as a base class for now.