Add releases to Artist
Each "root" entity is only going to contain references to the other
entities instead of the full entity. I'm doing this because otherwise
there'd be circular references, where a release can contain an artist,
and an artist can contain a release.

Maybe not the best, but I think it'll work. I _might_ add some helper
methods on each to get the full entities, but I am not sure how I want
to structure that stuff yet.